Mainly cleared up some sections with methods that allowed for better readability and less repetition in code.
For example, instead of registering listener by listener with the
`Bukkit.getPluginManager().registerEvents(new blah blah blah)`,
I took advantage of using the Arrays utility class to take advantage of the `.asList()` method along with a forEach for simpler registering of listeners. Used this method as well for registering SubCommands.

Changed the naming of some of the more commonly used methods so that they can follow java naming conventions.

A long with other changes, mainly introduced to improve the quality and readability of the code.

Fixed a common issue that involved using the == operator instead of the .equals() method when comparing strings.